Separate default implementaion
