+++ core/src/Kokkos_HostSpace.hpp
@@ -38,7 +38,6 @@ static_assert(false,
 
 #include "impl/Kokkos_HostSpace_deepcopy.hpp"
 #include <impl/Kokkos_MemorySpace.hpp>
-#include <impl/Kokkos_ZeroMemset_fwd.hpp>
 
 /*--------------------------------------------------------------------------*/
 
@@ -241,27 +240,6 @@ class SharedAllocationRecord<Kokkos::HostSpace, void>
 namespace Kokkos {
 
 namespace Impl {
-
-template <class DT, class... DP>
-struct ZeroMemset<typename HostSpace::execution_space, DT, DP...> {
-  ZeroMemset(const typename HostSpace::execution_space& exec,
-             const View<DT, DP...>& dst,
-             typename View<DT, DP...>::const_value_type&) {
-    // Host spaces, except for HPX, are synchronous and we need to fence for HPX
-    // since we can't properly enqueue a std::memset otherwise.
-    // We can't use exec.fence() directly since we don't have a full definition
-    // of HostSpace here.
-    hostspace_fence(exec);
-    using ValueType = typename View<DT, DP...>::value_type;
-    std::memset(dst.data(), 0, sizeof(ValueType) * dst.size());
-  }
-
-  ZeroMemset(const View<DT, DP...>& dst,
-             typename View<DT, DP...>::const_value_type&) {
-    using ValueType = typename View<DT, DP...>::value_type;
-    std::memset(dst.data(), 0, sizeof(ValueType) * dst.size());
-  }
-};
 
 template <>
 struct DeepCopy<HostSpace, HostSpace, DefaultHostExecutionSpace> {
