+++ core/src/impl/Kokkos_ViewMapping.hpp
@@ -32,7 +32,7 @@
 #include <impl/Kokkos_Atomic_View.hpp>
 #include <impl/Kokkos_Tools.hpp>
 #include <impl/Kokkos_StringManipulation.hpp>
-#include <impl/Kokkos_ZeroMemset_fwd.hpp>
+#include <impl/Kokkos_ZeroMemset.hpp>
 
 //----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
